ml-models/requirements.txt
numpy==1.22.3
pandas==1.4.2
scikit-learn==1.1.1
imbalanced-learn==0.9.1
pyyaml==6.0
python-json-logger==2.0.4